backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.9.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
python-multipart==0.0.12
pandas==2.2.3
scikit-learn==1.5.2
numpy==2.1.1
python-dotenv==1.0.1
email-validator==2.2.0
